Refactor code

+++ src/main/java/com/example/service/CategoryService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface CategoryService {
     List<Category> getAll();
 
-    Category getById(int id);
+    Category getById(long id);
 
     Category getByName(String name);
